% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prize_winners_by_keyword.R
\name{get_prize_years_by_keyword}
\alias{get_prize_years_by_keyword}
\title{Find winning years by keyword}
\usage{
get_prize_years_by_keyword(keyword)
}
\arguments{
\item{keyword}{Single word or phrase to search in the award rationales}
}
\value{
A vector of years which mention the keyword
}
\description{
Find winning years by keyword
}
\examples{
get_prize_years_by_keyword("therapy")
}
