% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_winner_by_year.R
\name{get_winner_by_year}
\alias{get_winner_by_year}
\title{Find people by the year they won}
\usage{
get_winner_by_year(years)
}
\arguments{
\item{years}{Vector of years between 1901 and present}
}
\value{
A tibble of winner details
}
\description{
Find people by the year they won
}
\examples{
get_winner_by_year(1960:1969)
}
